% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CrSolve}
\alias{CrSolve}
\title{Function providing the arterial blood concentration measured during the scan for one-tissue compartmental model.}
\usage{
CrSolve(t)
}
\arguments{
\item{t}{time}
}
\value{
A value representing Ca(t)
}
\description{
This function outputs the arterial blood concentration of the tracer measured during the scan.
}
\keyword{BayesPET}
