% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpntPET.R
\name{GenCurve}
\alias{GenCurve}
\title{Generate data from lp-ntPET model}
\usage{
GenCurve(Ct, Cr, Ti, R1, K2, K2a, gamma, tD, tP, alpha)
}
\arguments{
\item{Ct}{vector of concentration values of the tracer in the target tissue.}

\item{Cr}{vector of concentration values in the reference regions.}

\item{Ti}{vector of time points of observations.}

\item{R1}{value of the R1 parameter.}

\item{K2}{value of the K2 parameter.}

\item{K2a}{value of the K2a parameter.}

\item{gamma}{value of the activation parameter.}

\item{tD}{delay time at which the response starts.}

\item{tP}{peak time of maximal response.}

\item{alpha}{sharpeness parameters of the response function.}
}
\value{
M1 data under the model with no activation (gamma=0).

M2 data under the model with activation (gamma different from 0).
}
\description{
This function generate the data from the lp-ntPET model as in Normandin (2012).
}
\keyword{PETabc}
