% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{inputfunction}
\alias{inputfunction}
\title{Input function}
\usage{
inputfunction(t, type)
}
\arguments{
\item{t}{time}

\item{type}{type of the model. If type=1 is Model 1 in Feng et al. Type=2 represents Model 2 in Feng et al.}
}
\value{
A vector of concentrations for each time point.
}
\description{
Input function, values from Feng et al. 1993, Models for computer simulation
studies of input functions for tracer kinetic modeling with positron
emission tomography.
}
\keyword{BayesPET}
