% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpntPET.R
\name{lp_ntPETabc}
\alias{lp_ntPETabc}
\title{PET ABC function for the lp-nt model.}
\usage{
lp_ntPETabc(
  Ct,
  Cr,
  Ti,
  S = 10^5,
  R1a = 0.8,
  R1b = 1.1,
  K2alpha = 0.2,
  K2beta = 0.6,
  K2a.alpha = 0.1,
  K2a.beta = 0.4,
  gamma.a = 0,
  gamma.b = 0.2,
  tD.a = 15,
  tD.b = 25,
  tP.b = 40,
  alpha.a = 0,
  alpha.b = 3,
  lsq = 1,
  PLOT = T
)
}
\arguments{
\item{Ct}{vector of concentration values of the tracer in the target tissue.}

\item{Cr}{vector of concentration values in the reference regions.}

\item{Ti}{vector of time points of observations.}

\item{S}{number of simulations. Default at 10^5.}

\item{R1a}{upper bound of the uniform prior distribution for R1. Default at 1.6.}

\item{K2alpha}{lower bound of the uniform prior distribution for K2. Default at 0.}

\item{K2beta}{upper bound of the uniform prior distribution for K2. Default at 1.}

\item{K2a.alpha}{lower bound of the uniform prior distribution for K2a. Default at 0.}

\item{K2a.beta}{upper bound of the uniform prior distribution for K2a. Default at 0.6.}

\item{gamma.a}{lower bound of the uniform prior distribution for gamma. Default at 0.}

\item{gamma.b}{upper bound of the uniform prior distribution for gamma. Default at 0.2.}

\item{tD.a}{lower bound of the uniform prior distribution for tD. Default at 18.}

\item{tD.b}{upper bound of the uniform prior distribution for tD. Default at 22.}

\item{tP.b}{upper bound of the uniform prior distribution for tP. Default at 40.}

\item{alpha.a}{lower bound of the uniform prior distribution for alpha. Default at 0.}

\item{alpha.b}{upper bound of the uniform prior distribution for alpha. Default at 3.}

\item{lsq}{type of least squares estimation: 1 for non-negative least squares, 2 for weighted non-negative least squares. Default at 1.}

\item{PLOT}{If plots have to be produced. Default at TRUE.}
}
\value{
ABCout_act a matrix with values simulated from the posterior distribution of the parameters of the model with activation.

Smat_act a matrix with values of summary statistics of the simulated curves for the model with activation.

ABCout_accepted_act a matrix with values simulated from the posterior distribution of the parameters
of the model with activation; it shows only the values accepted according to the automatically selected threshold.

error_act vector of computed squared differences among the observed and the simulated summary statistics for the model with activation.

tol_act automatically selected tolerance level for the model with activation; this tolerance level is used to define the matrix ABCout_accepted.

ABCout_noact a matrix with values simulated from the posterior distribution of the parameters of the model with no activation.

Smat_noact a matrix with values of summary statistics of the simulated curves for the model with no activation.

ABCout_accepted_noact a matrix with values simulated from the posterior distribution of the parameters
of the model with no activation; it shows only the values accepted according to the automatically selected threshold.

error_noact vector of computed squared differences among the observed and the simulated summary statistics for the model with no activation.

tol_noact automatically selected tolerance level for the model with no activation; this tolerance level is used to define the matrix ABCout_accepted.
}
\description{
This function peforms ABC inference for the lp-nt model.
}
\keyword{PETabc}
