% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelchoice_abc.R
\name{MC_abc}
\alias{MC_abc}
\title{Model choice function.}
\usage{
MC_abc(
  y,
  tspan,
  abc_single,
  abc_two,
  tol1 = NULL,
  tol2 = NULL,
  inputfunction. = inputfunction,
  type = 2,
  PLOT = F
)
}
\arguments{
\item{y}{vector of measured radioactivity concentrations in a voxel for
each time-point. This can be in terms of observations or directly in terms
of voxel time activity curve.}

\item{tspan}{vector of time points of observations.}

\item{abc_single}{output of the PETabc() function for the single-tissue
single-tissue compartment model}

\item{abc_two}{output of the PETabc() function for the two-tissue
single-tissue compartment model}

\item{tol1}{tolerance level for the single-tissue model.
If tol1=NULL, the tolerance level automatically chosen by the PETabc()
function will be used. Default NULL.}

\item{tol2}{tolerance level for the two-tissue model.
If tol2=NULL, the tolerance level automatically chosen by the PETabc()
function will be used. Default NULL.}

\item{type}{The type of model in the input function.}

\item{PLOT}{If plots have to be produced. Default at FALSE.}

\item{inputfunction}{A function describing the input function.
Default inputfunction().}
}
\value{
postProb posterior probabilities of the single-tissue compartment model.

RMSEsingle root mean squared error of the single-tissue compartment model (on a grid of tolerance levels).

RMSEtwo root mean squared error of the two-tissue compartment model (on a grid of tolerance levels).
}
\description{
This function peforms model choice between the single-tissue and the
two-tissue models based on the outputs of the PETabc() function.
}
\keyword{PETabc}
