% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelchoice_lpntPET.R
\name{MC_abc_lpntPET}
\alias{MC_abc_lpntPET}
\title{Model choice function for the lp-ntPET model.}
\usage{
MC_abc_lpntPET(Ct, Cr, Ti, abc_out, tol1 = NULL, tol2 = NULL, PLOT = F)
}
\arguments{
\item{Ct}{vector of concentration values of the tracer in the target tissue.}

\item{Cr}{vector of concentration values in the reference regions.}

\item{Ti}{vector of time points of observations.}

\item{abc_out}{output of the lp_ntPETabc() function for the lp-ntPET model (both with and without activation).}

\item{tol1}{tolerance level for the lp-ntPET model with no activation. The tolerance is given in terms
of percentage of posterior values to keep.
If tol1=NULL, the tolerance level automatically chosen by the lp_ntPETabc()
function will be used. Default NULL.}

\item{tol2}{tolerance level for the lp-ntPET model with activation. The tolerance is given in terms
of percentage of posterior values to keep.
If tol2=NULL, the tolerance level automatically chosen by the lp_ntPETabc()
function will be used. Default NULL.}

\item{PLOT}{If plots have to be produced. Default at FALSE.}
}
\value{
postProb posterior probabilities of the lp-ntPET model with no activation.

RMSEnoact root mean squared error of the lp-ntPET model with no activation (on a grid of tolerance levels).

RMSEact root mean squared error of the lp-ntPET model with activation (on a grid of tolerance levels).
}
\description{
This function peforms model choice between the lp-ntPET model with activation and without activation.
}
\keyword{PETabc}
