% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelchoice_lpntPET.R
\name{mode_value}
\alias{mode_value}
\title{Mode function.}
\usage{
mode_value(xvec)
}
\arguments{
\item{xvec}{vector of values for which to compute the kernel density estimate.}
}
\value{
The modal value of the kernel density estimate of the vector used as input.
}
\description{
This function computes the modal value of a kernel density estimate of a vector of
(possibly) continuous values.
}
\keyword{PETabc}
