% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{oneTissueTAC}
\alias{oneTissueTAC}
\title{Function solving the differential equation for the one-tissue compartmental model}
\usage{
oneTissueTAC(tspan, K1, k2)
}
\arguments{
\item{tspan}{vector with the times}

\item{K1}{constant for the flux from the blood to the tissue.}

\item{k2}{constant for the flux from the tissue to the blood.}
}
\value{
A value representing the estimated C(t)
}
\description{
This function solves the ordinary differential equation describing the net tracer flux
into tissue.
}
\keyword{BayesPET}
