% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abc.R
\name{PETabc}
\alias{PETabc}
\title{PET ABC function}
\usage{
PETabc(
  y,
  t,
  N = 1e+05,
  inputfunction. = inputfunction,
  type = 2,
  a1 = 0,
  b1 = 0.2,
  a2 = 0.3,
  b2 = 0.5,
  a3 = 0,
  b3 = 0.1,
  a4 = 0,
  b4 = 0.2,
  tol
)
}
\arguments{
\item{y}{vector of measured radioactivity concentrations in a voxel for each time-point. This can be in terms of observations or directly in terms of voxel time activity curve.}

\item{t}{vector of time points of observations.}

\item{N}{Number of simulations. Default to 10^5.}

\item{type}{The type of model in the input function.}

\item{a1}{Minimum of the uniform prior for K1. Default to 0 .}

\item{b1}{Maximum of the uniform prior for K1. Default to 0.2 .}

\item{a2}{Minimum of the uniform prior for K2. Default to 0.3 .}

\item{b2}{Maximum of the uniform prior for K2. Default to 0.5 .}

\item{a3}{Minimum of the uniform prior for K3. Default to 0 .}

\item{b3}{Maximum of the uniform prior for K3. Default to 0.1 .}

\item{a4}{Minimum of the uniform prior for K4. Default to 0 .}

\item{b4}{Maximum of the uniform prior for K4. Default to 0.2 .}

\item{tol}{Tolerance level of the abc function.}

\item{inputfunction}{A function describing the input function. Default inputfunction().}
}
\value{
ABCres_single an object of class abc for the single-tissue compartment model.

ABCres_two an object of class abc for the two-tissue compartment model.
}
\description{
This function peforms ABC inference for the compartimental models.
}
\keyword{PETabc}
