% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpntPET.R
\name{rep.row}
\alias{rep.row}
\title{Repeating rows}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{vector to be repeated. The length of this vector will be the number of columns
of the matrix.}

\item{n}{scalar. It indicates the number of times the vector is repeated. It will be
the number of rows of the matrix.}
}
\value{
A matrix with dimension n x length(x)
}
\description{
This function repeats the same vector to create a matrix with n repeated rows.
}
