% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{singtissue}
\alias{singtissue}
\title{Single-tissue compartment model}
\usage{
singtissue(t, yobs, parms, inputfunction. = inputfunction, type = 2)
}
\arguments{
\item{t}{time}

\item{parms}{vector of two elements: K1, constant for the flux from blood
to tissue, k2, constant for the flux from tissue to blood.}

\item{type}{type of the input function.}

\item{y}{a matrix or dataframe of measured radioactivity concentration in a voxel
for each time-point. This can be in terms of observations or
directly in terms of voxel time activity curve. The first column refers
to the observations from the tissue.}

\item{inputfunction}{input function. Default inputfunction()}
}
\value{
A list with the output of the differential equation.
}
\description{
This function produce the output of the differential equation describing the
concentration in the reference tissue input function. The model considers a
bidirectional flux between tissue and blood.
}
\keyword{BayesPET}
