% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{TAC_2_Compartment}
\alias{TAC_2_Compartment}
\title{Simulation for two-tissue compartment model}
\usage{
TAC_2_Compartment(
  K1 = 0.0918,
  k2 = 0.4484,
  k3 = 0.0482,
  k4 = 0.1363,
  noise = "Pois",
  PLOT = F,
  l2 = 1,
  l1 = 3,
  tspan = 1:60,
  inputfunction. = inputfunction,
  type = 2
)
}
\arguments{
\item{K1}{constant for the flux from the blood to the tissue. Default to 0.0918.}

\item{k2}{constant for the flux from the tissue to the blood. Default to 0.4484.}

\item{k3}{constant for the flux from the first to the second tissue. Default to 0.0482.}

\item{k4}{constant for the flux from the second to the first tissue. Default to 0.1363.}

\item{noise}{type of error noise. Default Poisson (Pois), altnernative Gaussian ("norm"), or Poisson-normal ("mixed")}

\item{PLOT}{if a plot has to be produced. Default FALSE.}

\item{l2}{noise level for Poisson model. Default decreasing 1:20.}

\item{l1}{noise level for the normal model. Default increasing 1:10.}

\item{tspan}{vector of time points. Default 1:60}

\item{type}{type of the input function.}

\item{inputfunction}{input function. Default inputfunction()}
}
\value{
A list with two vectors with different type of noise.
}
\description{
Time Activity Curves Simulations with Normal distributed and Poisson noise for
the two-tissue compartment model.
}
\keyword{BayesPET}
