% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{twotissue}
\alias{twotissue}
\title{Two-tissue compartment model}
\usage{
twotissue(t, yobs, parms, inputfunction. = inputfunction, type = 2)
}
\arguments{
\item{t}{time}

\item{yobs}{matrix or dataframe of measured radioactivity concentration in a
voxel for each time-point. This can be in terms of observations or
directly in terms of voxel time activity curve. The
first column refers to the observations from the
first tissue, the second column refers to the observations from the
second tissue.}

\item{parms}{vector of three elements: K1, constant for the flux from blood to
tissue, k2, constant for the flux from tissue to blood, k3, constant
for the flux from the first tissue to the second, k4, constant for the flux
from the second to the first tissue.}

\item{type}{type of the input function.}

\item{inputfunction}{input function. Default inputfunction()}
}
\value{
A list with the output of the differential equation
}
\description{
This function produce the output of the differential equation describing the
concentration in a two-tissue compartment model. First compartment is for the input function,
the concentration of tracer in plasma or blood curve as a function of time C0(t).
The following two compartments are for the two distinct kinetic compartments
in the tissue (C1(t) and C2(t)), representing, for instance, the free
and receptor-bound tracer concentrations.
Compartments are connected with four rate constants, K1, k2, k3, and k4.
}
\keyword{BayesPET}
